requirements/requirement.txt
qiniu==7.2.4
tornado==6.0.2
Twisted==19.2.0
requests==2.22.0
SQLAlchemy==1.3.4
supervisor==4.0.3

six
pyglet==1.3.2
splinter==0.10.0

redis==3.2.1
pymongo==3.8.0
PyMySQL==0.9.3
MySQL-python==1.2.5

Faker==1.0.7
python-testdata==1.0.5
python-dotenv==0.10.2

git+https://github.com/wujuguang/autoloads.git
